backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "mariadb>=1.1.14",
    "pandas>=3.0.2",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
    "pymongo>=4.17.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.24",
    "uvicorn>=0.41.0",
]
